Removed simulation directory

+++ scripts/newprofiling.sh
@@ -6,56 +6,68 @@ if [ -z $chartdays ]
 then
  chartdays=600
 fi
-datadir=$5
-if [ -z $datadir ]
+tmpdir=$5
+if [ -z $tmpdir ]
 then
- datadir=/z/data
+ tmpdir=/z/data
 fi
+indir=$6
+mpvdir=$indir/mpv
 
-simdir=${datadir}/mpv/simulation
-sigdir=${simdir}/signals
-syndir=${simdir}/synopsis
-prfdir=${simdir}/profiling
-patdir=${simdir}/patterns
+sigdir=${mpvdir}/signals
+syndir=${mpvdir}/synopsis
+prfdir=${mpvdir}/profiling
+patdir=${mpvdir}/patterns
 
 if [ $opt -lt 3 ]
 then
  if [ $opt -eq 1 ]
  then
-  params="-ps -C"
+  params="-psj 1 -C"
  else
   params="-ps -D s -C"
  fi
  if ! test -d ${prfdir}/$counter
  then
-  mkdir ${prfdir}/$counter
+  mkdir -p ${prfdir}/$counter
  fi
- > $logfile
  logfile=${prfdir}/$counter/$counter.log
- rm ${prfdir}/$counter/*.png | tee -a $logfile
+ > $logfile
+ if ! [ $opt -eq 1 ]
+ then
+  rm ${prfdir}/$counter/*.png | tee -a $logfile
+ fi
 else
- params="-ps -D p -C"
+ params="-psj2 -Dp -C"
  if ! test -d ${patdir}/$counter
  then
-  mkdir ${patdir}/$counter
+  mkdir -p ${patdir}/$counter
  fi
  logfile=${patdir}/$counter/$counter.log
  > $logfile
  rm ${patdir}/$counter/*.png | tee -a $logfile
 fi
-> $logfile
-> ${simdir}/signals/$counter-signals.csv
-rm ${syndir}/${counter}-*.png | tee -a $logfile
-rm ${sigdir}/${counter}-signals.csv.* | tee -a $logfile
-rm ${sigdir}/${counter}-signals.csv | tee -a $logfile
-
-python analytics/mvpchart.py $counter $params -S $dates -c $chartdays -e ${datadir} | tee -a $logfile
-
-if [ $opt -lt 3 ]
+if ! [ $opt -eq 1 ]
 then
- mv ${simdir}/synopsis/$counter-*.png ${prfdir}/$counter/
-else
- mv ${simdir}/synopsis/$counter-*.png ${patdir}/$counter/
+ > ${sigdir}/$counter-signals.csv
+ rm ${syndir}/${counter}-*.png | tee -a $logfile
+ rm ${tmpdir}/mpv/signals/${counter}-signals.csv.* | tee -a $logfile
+ rm ${tmpdir}/mpv/signals/${counter}-signals.csv | tee -a $logfile
 fi
-#cp ${simdir}/signals/$counter-signals.csv ${datadir}/mpv/signals/
-uniq ${simdir}/signals/$counter-signals.csv > ${datadir}/mpv/signals/${counter}-signals.csv
+
+python analytics/mvpchart.py $counter $params -S $dates -c $chartdays -e ${indir} | tee -a $logfile
+
+if [ $opt -eq 1 ]
+then
+ #cp $tmpdir/json/$counter.json $indir/json
+ tar czvf $indir/json/$counter.tgz $tmpdir/json/$counter.*.json
+else
+ if [ $opt -lt 3 ]
+ then
+  mv ${tmpdir}/mpv/synopsis/$counter-*.png ${prfdir}/$counter/
+ else
+  mv ${tmpdir}/mpv/synopsis/$counter-*.png ${patdir}/$counter/
+ fi
+ #cp ${tmpdir}/mpv/signals/$counter-signals.csv ${sigdir}/
+ uniq ${tmpdir}/mpv/signals/$counter-signals.csv > ${sigdir}/${counter}-signals.csv
+fi
